+++ packages/package.go
@@ -50,6 +50,8 @@ type Package struct {
 
 	versionSemVer *semver.Version
 
+	specMajorMinorSemVer *semver.Version
+
 	fsBuilder FileSystemBuilder
 	resolver  RemoteResolver
 }
@@ -319,15 +321,33 @@ func (p *Package) setRuntimeFields() error {
 
 	p.versionSemVer, err = semver.StrictNewVersion(p.Version)
 	if err != nil {
-		return errors.Wrap(err, "invalid package version")
+		return fmt.Errorf("invalid package version: %w", err)
 	}
 
 	if p.Conditions != nil && p.Conditions.Kibana != nil {
 		p.Conditions.Kibana.constraint, err = semver.NewConstraint(p.Conditions.Kibana.Version)
 		if err != nil {
-			return errors.Wrapf(err, "invalid Kibana versions range: %s", p.Conditions.Kibana.Version)
+			return fmt.Errorf("invalid Kibana versions range %s: %w", p.Conditions.Kibana.Version, err)
 		}
 	}
+
+	// Packages from proxy mode do not have "format_version" field
+	if p.FormatVersion == "" {
+		return nil
+	}
+
+	specSemVer, err := semver.StrictNewVersion(p.FormatVersion)
+	if err != nil {
+		return fmt.Errorf("invalid format spec version '%s': %w", p.FormatVersion, err)
+	}
+
+	specMajorMinorVersion := fmt.Sprintf("%d.%d.0", specSemVer.Major(), specSemVer.Minor())
+
+	p.specMajorMinorSemVer, err = semver.StrictNewVersion(specMajorMinorVersion)
+	if err != nil {
+		return fmt.Errorf("invalid format spec version '%s': %w", specMajorMinorVersion, err)
+	}
+
 	return nil
 }
 
@@ -399,6 +419,28 @@ func (p *Package) WorksWithCapabilities(capabilities []string) bool {
 		}
 	}
 	return true
+}
+
+func (p *Package) HasCompatibleSpec(specMin, specMax, kibanaVersion *semver.Version) (bool, error) {
+	if specMin == nil && specMax == nil {
+		return true, nil
+	}
+
+	constraints := []string{}
+	if specMin != nil {
+		constraints = append(constraints, fmt.Sprintf(">=%s", specMin.String()))
+	}
+	if specMax != nil {
+		constraints = append(constraints, fmt.Sprintf("<=%s", specMax.String()))
+	}
+
+	fullConstraint := strings.Join(constraints, ",")
+	constraint, err := semver.NewConstraint(fullConstraint)
+	if err != nil {
+		return false, fmt.Errorf("cannot create constraint %s: %w", fullConstraint, err)
+	}
+
+	return constraint.Check(p.specMajorMinorSemVer), nil
 }
 
 func (p *Package) IsNewerOrEqual(pp *Package) bool {