+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add new query parameter "capabilities" in search endpoint [#1054](https://github.com/elastic/package-registry/pull/1054)
 * Add new query parameter "capabilities" in categories endpoint [#1061](https://github.com/elastic/package-registry/pull/1061))
+* Add new query parameters "spec.min" and "spec.max" in search endpoint [#1058](https://github.com/elastic/package-registry/pull/1058)
 
 ### Deprecated
 
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * New Security subcategory "Advanced Analytics (UEBA)" [#997](https://github.com/elastic/package-registry/pull/997)
+
 ### Deprecated
 
 ### Known Issues